Add all baggy instrumentation except for slow path
Since baggy bounds checking will change the control flow of the graph, we
need to use a function pass instead of a bsic block pass.  The function pass
loops through all the basic blocks in the function and if it finds any pointer
arithmetic in the block, splits up the block and inserts a baggy bounds
checking block in the middle.  The only thing left to do is to write the
slowPath function and include a branch to it at the end of the baggy bounds
block.