Mask off allocator metadata in size table entries
Only the lowest five bits in each size table entry are used for baggy bounds
checking because an allocation cannot be larger than 4 GB (32 bits) and log(32)
is 5.  The buddy allocator will use the remaining three bits to store metadata
about the allocated blocks.  We need to mask off these bits to ensure that they
don't interfere with the instrumented code.