Replace getelementptr instructions with PHI nodes
This solves the problem of updating all instructions following a getelementptr
instruction to refer to the PHI node created by the baggy bounds
instrumentation.  We first clone the getelementptr instruction and then replace
it with a PHI node.  We later insert the cloned instruction into the baggy
baggy bounds instrumentation block and update the PHI node to add the incoming
paths from the normal and slowPath branches.